Google maps. Search almost every thing near you. Nearby locations.
NearMe application displays nearby locations in 5 to 20 KM radius of
user's current location on startup. Following are the main features of
NearMe application:
1. User can change place type e.g. hospital, school, university etc by
tapping on place type displayed in the top bar of application. After
tapping on top bar it displays a list of type of places. By selecting
any type application starts fetching places and after that it displays
them on the map in the form of markers and in the form of list below the
map.
2. User can change his radius distance e.g. 5 KM, 10 KM, 15 KM, 20 KM
etc from his current location by tapping on 'KM' option displayed in the
top bar. After tapping on top bar application again fetches the places
in the selected radius and displays them on the map in the form of
markers and in the form of list below the map.

3.User can find details about any place either by looking in the list
shown below the map or by tapping any desired marker displayed on the
map. After tapping on the marker a popup will be displayed showing the
name, distance , address of the place and two buttons i.e 'CANCEL',
'PATH'.

4. User can find the path from his current location to the desired place
either by tapping on 'PATH' button displayed in the popup against marker
or by tapping on any list item shown below the map.